_example: change structure to have one example per directory
