--- _example/default.go
@@ -1,36 +1,0 @@
-package main
-
-import (
-	"log"
-	"math/rand"
-	"net/http"
-	"strconv"
-	"time"
-
-	"github.com/arl/statsviz"
-)
-
-func main() {
-	// Force the GC to work to make the plots "move".
-	go work()
-
-	// Register statsviz handlers on the default serve mux.
-	statsviz.RegisterDefault()
-	log.Fatal(http.ListenAndServe(":8080", nil))
-}
-
-func work() {
-	// Generate some allocations
-	m := map[string][]byte{}
-
-	for {
-		b := make([]byte, 512+rand.Intn(16*1024))
-		m[strconv.Itoa(len(m)%(10*100))] = b
-
-		if len(m)%(10*100) == 0 {
-			m = make(map[string][]byte)
-		}
-
-		time.Sleep(10 * time.Millisecond)
-	}
-}
