+++ _example/gorilla/main.go
@@ -1,18 +1,17 @@
 package main
 
 import (
-	"math/rand"
 	"net/http"
-	"strconv"
-	"time"
+
+	"github.com/gorilla/mux"
 
 	"github.com/arl/statsviz"
-	"github.com/gorilla/mux"
+	example "github.com/arl/statsviz/_example"
 )
 
 func main() {
 	// Force the GC to work to make the plots "move".
-	go work()
+	go example.Work()
 
 	// Create a Gorilla router and register statsviz handlers.
 	r := mux.NewRouter()
@@ -22,20 +21,4 @@ func main() {
 	mux := http.NewServeMux()
 	mux.Handle("/", r)
 	http.ListenAndServe(":8080", mux)
-}
-
-func work() {
-	// Generate some allocations
-	m := map[string][]byte{}
-
-	for {
-		b := make([]byte, 512+rand.Intn(16*1024))
-		m[strconv.Itoa(len(m)%(10*100))] = b
-
-		if len(m)%(10*100) == 0 {
-			m = make(map[string][]byte)
-		}
-
-		time.Sleep(10 * time.Millisecond)
-	}
 }
