+++ _example/https/main.go
@@ -2,22 +2,20 @@ package main
 
 import (
 	"log"
-	"math/rand"
 	"net/http"
-	"strconv"
-	"time"
 
 	"github.com/arl/statsviz"
+	example "github.com/arl/statsviz/_example"
 )
 
 func main() {
 	// Force the GC to work to make the plots "move".
-	go work()
+	go example.Work()
 
 	const (
 		// Use your own certificates and key files.
-		certFile = "_example/cert.pem"
-		keyFile  = "_example/key.pem"
+		certFile = "_example/https/cert.pem"
+		keyFile  = "_example/https/key.pem"
 	)
 
 	// Create a serve mux and register statsviz handlers.
@@ -27,20 +25,4 @@ func main() {
 	}
 
 	log.Fatal(http.ListenAndServeTLS(":8080", certFile, keyFile, mux))
-}
-
-func work() {
-	// Generate some allocations
-	m := map[string][]byte{}
-
-	for {
-		b := make([]byte, 512+rand.Intn(16*1024))
-		m[strconv.Itoa(len(m)%(10*100))] = b
-
-		if len(m)%(10*100) == 0 {
-			m = make(map[string][]byte)
-		}
-
-		time.Sleep(10 * time.Millisecond)
-	}
 }
