+++ _example/mux/main.go
@@ -2,17 +2,15 @@ package main
 
 import (
 	"log"
-	"math/rand"
 	"net/http"
-	"strconv"
-	"time"
 
 	"github.com/arl/statsviz"
+	example "github.com/arl/statsviz/_example"
 )
 
 func main() {
 	// Force the GC to work to make the plots "move".
-	go work()
+	go example.Work()
 
 	// Create a serve mux and register statsviz handlers.
 	mux := http.NewServeMux()
@@ -21,20 +19,4 @@ func main() {
 	}
 
 	log.Fatal(http.ListenAndServe(":8080", mux))
-}
-
-func work() {
-	// Generate some allocations
-	m := map[string][]byte{}
-
-	for {
-		b := make([]byte, 512+rand.Intn(16*1024))
-		m[strconv.Itoa(len(m)%(10*100))] = b
-
-		if len(m)%(10*100) == 0 {
-			m = make(map[string][]byte)
-		}
-
-		time.Sleep(10 * time.Millisecond)
-	}
 }
